requirements.txt
aiogram>=3.0
python-docx>=1.1
reportlab>=4.0
pillow>=10.0
pypdf>=4.0
